refactor(terminal): consolidate duplicate icon rendering logic
- Create shared TerminalIcon component in Terminal/
- Remove duplicate getTerminalIcon functions from 3 files
- Preserve size overrides via className (w-3, w-3.5, w-4)
- Fix missing brandColor in TerminalListItem
- Export TerminalIcon from Terminal index
- Reduce codebase by 60 lines
