+++ src/components/Terminal/index.ts
@@ -12,3 +12,5 @@ export { StateBadge } from "./StateBadge";
 export { ActivityBadge } from "./ActivityBadge";
 export type { ActivityStatus, ActivityType } from "./ActivityBadge";
 export { TerminalCountWarning, SOFT_TERMINAL_LIMIT } from "./TerminalCountWarning";
+export { TerminalIcon } from "./TerminalIcon";
+export type { TerminalIconProps } from "./TerminalIcon";
